Update lnm_free_item implementation to prevent recursion
The previous implementation of the function used recursion to walk
down the log tree of an Event item. This new implementation makes use
of a breadcrumb-style navigation array and some loop logic to walk
the log tree without the use of recursion.

This is particularly helpful to prevent stack overflows. While a
stack overflow would likely never occur under normal circumstances,
there are still a couple issues with using regular recursion:

 - The argument supplied to the function is created by the developer,
   not the library. Depending on what the developer does, the log
   tree could be tens or hundreds of items deep. While it should
   never be this way, any developer who does this may experience a
   stack overflow.

 - Even if the log tree is shallow, a stack overflow is still
   possible on embedded systems with limited stack sizes.

Because of these limitations, it's more ideal to completely
eliminate the use of recursion rather than ask the developer to
reserve more stack space himself/herself.