Add -std=c11 specificity to gcc calls
Specifies the version of the C standard to use for compilation. C11 was
chosen because it's the earliest version of the C standard that compiles
the library.

C99 failed compilation due to the use of forward-declared structs in the
same scope of the translation unit (disallowed by the C99 standard), as
well as the requirement of the _Noreturn function declaration keyword
(not introduced until C11).

Until this point, the library has compiled with C18, and, along with any
newer versions of the C standard that come out, should continue to as
long as it works with C11.